+++ plugin/src/main/java/com/stratio/cassandra/lucene/search/condition/BooleanCondition.java
@@ -17,10 +17,11 @@
 package com.stratio.cassandra.lucene.search.condition;
 
 import com.google.common.base.Objects;
-import com.stratio.cassandra.lucene.IndexException;
 import com.stratio.cassandra.lucene.schema.Schema;
+import com.stratio.cassandra.lucene.util.Log;
 import org.apache.lucene.search.BooleanClause.Occur;
 import org.apache.lucene.search.BooleanQuery;
+import org.apache.lucene.search.MatchAllDocsQuery;
 import org.apache.lucene.search.Query;
 
 import java.util.LinkedList;
@@ -59,13 +60,6 @@ public class BooleanCondition extends Condition {
         this.must = must == null ? new LinkedList<Condition>() : must;
         this.should = should == null ? new LinkedList<Condition>() : should;
         this.not = not == null ? new LinkedList<Condition>() : not;
-
-        if (this.must.isEmpty() && this.should.isEmpty() && !this.not.isEmpty()) {
-            throw new IndexException("Invalid Boolean Query: lucene does not accept pure NOT queries: " +
-                                     "http://lucene.apache.org/core/2_9_4/queryparsersyntax.html#NOT " +
-                                     "add a 'must : [{ type : \"all\" }]' in your query");
-        }
-
     }
 
     /** {@inheritDoc} */
@@ -81,6 +75,10 @@ public class BooleanCondition extends Condition {
         }
         for (Condition query : not) {
             luceneQuery.add(query.query(schema), Occur.MUST_NOT);
+        }
+        if (must.isEmpty() && should.isEmpty() && !not.isEmpty()) {
+            Log.warn("Performing resource-intensive pure negation search");
+            luceneQuery.add(new MatchAllDocsQuery(), Occur.FILTER);
         }
         return luceneQuery;
     }